refactor(mcp): extract tool handler registry shared by MCP and agent
The 25 authoring tools were registered inline in mcp/server.mjs, so the
only way to run one was to speak MCP to a running server. Move them into
mcp/tool-handlers.mjs as a registry: createToolHandlers(deps) returns one
entry per tool carrying its name, title, description, input schema, safety
annotations, live routing flag and handler, and server.mjs registers from
that array through the same registerTool wrapper as before.

This is a pure move. The wrapper still owns the MCP-facing behaviour it
always did (the missing-annotations guard, the workspace_handle schema and
description augmentation, and LiveHub.runExclusive routing), so tool names,
order, schemas, annotations, error messages and live routing are unchanged
byte for byte over the protocol.

The registry owns the authoring state and exposes liveHub as a live binding
the owner installs with setLiveHub, which is what lets the handler bodies
move without touching a single reference. Motion job delivery stays in
server.mjs: the live hub's reconnect and cancel paths publish jobs too, so
generate_motion takes the registry and publisher as dependencies.

Importing mcp/tool-handlers.mjs has no side effects — no port, no chdir, no
signal handler, no artifact sweep — so an in-process agent can import it
without starting an MCP server. test/verify-tool-handlers.mjs asserts that,
the 25 names in order with their annotations, and that describe_scene called
directly returns exactly what the MCP server returned before the extraction.

Refs #127
