+++ tools/run-tests.mjs
@@ -123,6 +123,7 @@ const NODE_FILES = [
 	"test/verify-telemetry-state.mjs",
 	"test/verify-timeline-camera.mjs",
 	"test/verify-timeline-shots.mjs",
+	"test/verify-tool-handlers.mjs",
 	"test/verify-trim.mjs",
 	"test/verify-update-check.mjs",
 	"test/verify-usd-camera.mjs",
